use wrapper function again to fix argument passing
While the regex iterations induce single arguments being passed to the
printf-macro functions and not all (as it was the case for the
implementation using recursion), no need was assumed anymore for the
template wrapper function that passed an empty string for non-trivial
arguments and the argument itself otherwise (because printf doesn't
accept non-trivial arguments with clang++). With protected/private copy
constructors however, the function invoking the variadic macro function
call doesn't compile at all, because the ... arguments are apparently
copied. This commit hence fixes this issue by reuse of the wrapper
function mentioned above.